backend/requirements.txt
Django==2.2.5
django-cors-headers==2.4.0
djangorestframework==3.9.4
